+++ webapp2/src/view/BookApp.vue
@@ -1,46 +1,12 @@
 <template>
   <div>
-    <v-row class="mx-3">
-    書籍を検索
-    </v-row>
-    <v-row>
-      <v-col cols="3" class="pt-1 pb-0">
-        <v-text-field
-        v-model="searchTitle"
-        solo
-        clearable
-        label="書籍名"
-        variant="outlined"
-        width="400">
-        </v-text-field>
-      </v-col>
-      <v-col cols="2" class="pt-1 pb-0">
-        <v-select
-          solo
-          v-model="searchGenre"
-          :items="formGenreList"
-          label="ジャンル"
-          item-text="genreName"
-          item-value="genreId"
-          width="100"
-          >
-        </v-select>
-      </v-col>
-      <v-col cols="4" class="pt-1 pb-0">
-        <v-btn
-          color="light-blue lighten-1"
-          class="mt-2 white--text"
-          @click="searchBook">
-          <v-icon left>mdi-folder-search</v-icon>検索
-        </v-btn>
-        <v-btn
-          color="blue darken-3"
-          class="mt-2 ml-2 white--text"
-          @click="searchClear">
-          <v-icon left>mdi-backspace</v-icon>検索条件クリア
-        </v-btn>
-      </v-col>
-    </v-row>
+    <!-- 検索機能部 -->
+    <Search
+      :formGenreList="formGenreList"
+      @searchBook='searchBook'
+      @searchClear='searchClear'
+    />
+    <!-- 新規登録機能部 -->
     <v-row>
       <v-col class="pt-0">
         <v-btn
@@ -50,230 +16,62 @@
         </v-btn>
       </v-col>
     </v-row>
-    <!-- 登録されている書籍情報を表示するリスト -->
-    <v-data-table
-      :headers="headers"
+    <!-- 書籍一覧表示部 -->
+    <List
       :items="items"
-      @click:row="showReview"
-      sort-by="bookId"
-    >
-      <template v-slot:item.actions="{ item }">
-        <v-btn
-          small
-          color="orange lighten-1"
-          class="mx-1 white--text"
-          @click.stop="openEditForm(item)">
-          <v-icon left>mdi-pencil</v-icon>編集
-        </v-btn>
-        <v-btn
-          small
-          color="red accent-3"
-          class="mx-1 white--text"
-          @click.stop="openDeleteDialog(item)">
-          <v-icon left>mdi-delete</v-icon>削除
-        </v-btn>
-      </template>
-    </v-data-table>
+      @openEditForm='openEditForm'
+      @reloadItem='reloadItem'
+      @notifyFailure='notifyFailure'
+    />
+    <!-- 入力フォーム部 -->
+    <Form
+      :showForm="showForm"
+      :isNewBook="isNewBook"
+      :formGenreList="formGenreList"
+      :selectedBookId="selectedBookId"
+      :selectedTitle="selectedTitle"
+      :selectedGenreId="selectedGenreId"
+      :selectedPurchaseDate="selectedPurchaseDate"
+      :selectedPurchaseName="selectedPurchaseName"
+      :selectedReview="selectedReview"
+      @reloadItem='reloadItem'
+      @closeForm='closeForm'
+      @notifyFailure='notifyFailure'
+    />
     <!-- DBアクセス時に表示するオーバーレイ -->
     <v-overlay :value="overlay">
       <v-progress-circular indeterminate size="64">
       </v-progress-circular>
     </v-overlay>
-    <!-- 書籍登録／編集時に表示する入力フォーム -->
-    <div id="overlay" v-show="showForm">
-      <div id="formWindow">
-        <v-card>
-          <v-card-title>
-            書籍登録／編集
-          </v-card-title>
-          <v-card-text>
-            <v-form ref="form">
-              <v-row>
-                <v-col class="py-0">
-                  <v-text-field
-                    v-model="formTitle"
-                    label="タイトル"
-                    required
-                  ></v-text-field>
-                </v-col>
-              </v-row>
-              <v-row>
-                <v-col class="py-0">
-                  <v-select
-                    v-model="selectedFormGenreList"
-                    :items="formGenreList"
-                    label="ジャンル"
-                    item-text="genreName"
-                    item-value="genreId"
-                    required
-                  ></v-select>
-                </v-col>
-              </v-row>
-              <v-row>
-                <v-col class="py-0">
-                  <v-menu
-                    ref="menu"
-                    v-model="pickerMenu"
-                    :close-on-content-click="false"
-                    :return-value.sync="pickerDate"
-                    transition="scale-transition"
-                    offset-y
-                    min-width="auto"
-                  >
-                    <template v-slot:activator="{ on, attrs }">
-                      <v-text-field
-                        v-model="pickerDate"
-                        label="購入日"
-                        prepend-icon="mdi-calendar-cursor"
-                        readonly
-                        v-bind="attrs"
-                        v-on="on"
-                      ></v-text-field>
-                    </template>
-                    <v-date-picker
-                      v-model="pickerDate"
-                      no-title
-                      scrollable
-                    >
-                      <v-spacer></v-spacer>
-                      <v-btn
-                        class="mr-2"
-                        color="primary"
-                        @click="$refs.menu.save(pickerDate)"
-                      >
-                        決定
-                      </v-btn>
-                      <v-btn
-                        class="mr-2"
-                        color="secondary"
-                        @click="pickerMenu = false"
-                      >
-                        取消
-                      </v-btn>
-                    </v-date-picker>
-                  </v-menu>
-                </v-col>
-              </v-row>
-              <v-row>
-                <v-col class="py-0">
-                  <v-text-field
-                    v-model="formPurchaseName"
-                    label="購入者"
-                    required
-                  ></v-text-field>
-                </v-col>
-              </v-row>
-              <v-row>
-                <v-col class="py-0">
-                  <v-textarea
-                    v-model="formReview"
-                    filled
-                    label="レビュー"
-                    auto-grow
-                    clearable
-                  ></v-textarea>
-                </v-col>
-              </v-row>
-              <v-row>
-                <v-col class="py-0" align="center">
-                  <v-btn
-                    class="mr-4"
-                    color="primary"
-                    @click="handlerInsertOrUpdate"
-                  >
-                    登録
-                  </v-btn>
-                  <v-btn
-                    color="secondary"
-                    @click="closeForm"
-                  >
-                    取消
-                  </v-btn>
-                </v-col>
-              </v-row>
-            </v-form>
-          </v-card-text>
-        </v-card>
-      </div>
-    </div>
-    <!-- 書籍削除時に表示する確認ダイアログ -->
-    <v-dialog v-model="dialogDelete" max-width="25%">
-      <v-card>
-        <v-card-title>{{ dialogTitle }}</v-card-title>
-        <v-card-actions>
-          <v-btn
-            color="red accent-3"
-            class="mx-auto white--text"
-            @click="deleteItemConfirm"
-          >削除
-          </v-btn>
-          <v-btn
-            color="secondary"
-            class="mx-auto"
-            @click="closeDelete"
-          >取消
-          </v-btn>
-        </v-card-actions>
-      </v-card>
-    </v-dialog>
-    <!-- 書籍レビューを表示するダイアログ -->
-    <v-dialog
-      v-model="dialogReview"
-      persistent
-      max-width="35%">
-      <v-card>
-        <v-card-title>{{ reviewTitle }}</v-card-title>
-        <v-card-text>
-          <v-row>
-            <v-textarea
-              v-model="formReview"
-              solo
-              dense
-              readonly
-            ></v-textarea>
-          </v-row>
-        </v-card-text>
-        <v-card-actions>
-          <v-btn
-            color="secondary"
-            class="mx-auto"
-            @click="closeReview"
-          >閉じる
-          </v-btn>
-        </v-card-actions>
-      </v-card>
-    </v-dialog>
   </div>
 </template>
 
 <script>
+import Search from '@/components/Search.vue'
+import List from '@/components/List.vue'
+import Form from '@/components/Form.vue'
 export default {
+  components: {
+    Search,
+    List,
+    Form
+  },
   data () {
     return {
-      headers: [
-        { text: 'タイトル', value: 'title', width: '50%' },
-        { text: 'ジャンル', value: 'genreName', width: '14%' },
-        { text: '購入日', value: 'purchaseDate', width: '10%' },
-        { text: '購入者', value: 'purchaseName', width: '10%' },
-        { text: '操作', value: 'actions', width: '16%', sortable: false }
-      ],
       items: [],
-      formTitle: '',
       formGenreList: [],
-      selectedFormGenreList: '',
-      pickerDate: new Date().toISOString().substr(0, 10),
-      pickerMenu: false,
-      formPurchaseName: '',
-      formReview: '',
       selectedBookId: '',
       selectedTitle: '',
+      selectedGenreId: '',
+      selectedPurchaseDate: '',
+      selectedPurchaseName: '',
+      selectedReview: '',
       searchTitle: '',
       searchGenre: '',
       isFilteredView: false,
+      isNewBook: false,
       overlay: false,
-      showForm: false,
-      dialogDelete: false,
-      dialogReview: false
+      showForm: false
     }
   },
   async created () {
@@ -281,14 +79,7 @@ export default {
     await this.initialLoad()
     this.overlay = false
   },
-  computed: {
-    dialogTitle () {
-      return `【${this.selectedTitle}】を削除しますか？`
-    },
-    reviewTitle () {
-      return `【${this.formTitle}】のレビュー`
-    }
-  },
+  computed: {},
   methods: {
     // ページロード時の読み込み処理
     async initialLoad () {
@@ -300,6 +91,17 @@ export default {
       })
       this.items = initialGet[0]
       this.formGenreList = initialGet[1]
+    },
+    // 書籍テーブルリロード
+    async reloadItem () {
+      this.overlay = true
+      try {
+        this.items = await this.readBookTableAll() // 書籍一覧更新
+      } catch (e) {
+        this.notifyFailure()
+      }
+      this.overlay = false
+      this.isNewBook = false
     },
     // GASを使用しジャンルテーブルを読み込み、Promise返却
     readGenreTable () {
@@ -328,134 +130,38 @@ export default {
           .readBookTableFiltering(this.searchTitle, this.searchGenre)
       })
     },
-    // GASを使用し書籍テーブルへ新規登録、Promise返却
-    insertNewItem () {
-      return new Promise((resolve, reject) => {
-        google.script.run // eslint-disable-line no-undef
-          .withSuccessHandler((result) => resolve(result))
-          .withFailureHandler((error) => reject(error))
-          .insertNewBook(
-            this.formTitle,
-            this.selectedFormGenreList,
-            this.pickerDate,
-            this.formPurchaseName,
-            this.formReview
-          )
-      })
-    },
-    // GASを使用し書籍テーブルから書籍レコード削除、Promise返却
-    deleteItem () {
-      return new Promise((resolve, reject) => {
-        google.script.run // eslint-disable-line no-undef
-          .withSuccessHandler((result) => resolve(result))
-          .withFailureHandler((error) => reject(error))
-          .deleteBook(this.selectedBookId)
-      })
-    },
-    // GASを使用し書籍テーブルを更新、Promise返却
-    updateItem () {
-      return new Promise((resolve, reject) => {
-        google.script.run // eslint-disable-line no-undef
-          .withSuccessHandler((result) => resolve(result))
-          .withFailureHandler((error) => reject(error))
-          .updateBook(
-            this.selectedBookId,
-            this.formTitle,
-            this.selectedFormGenreList,
-            this.pickerDate,
-            this.formPurchaseName,
-            this.formReview
-          )
-      })
-    },
     // 書籍新規登録フォーム表示
     openForm () {
-      this.selectedBookId = 0
+      this.isNewBook = true
       this.showForm = true
     },
-    // 書籍新規登録サブミット
-    async submitInsertNewItem () {
-      this.overlay = true
-      try {
-        await this.insertNewItem() // 書籍登録
-        this.items = await this.readBookTableAll() // 書籍一覧更新
-      } catch (e) {
-        this.notifyFailure()
-      }
-      this.overlay = false
-      this.closeForm()
+    // 書籍編集フォーム表示
+    openEditForm (
+      bookId,
+      title,
+      genreId,
+      purchaseDate,
+      purchaseName,
+      review) {
+      this.selectedBookId = bookId
+      this.selectedTitle = title
+      this.selectedGenreId = genreId
+      this.selectedPurchaseDate = purchaseDate
+      this.selectedPurchaseName = purchaseName
+      this.selectedReview = review
+      this.isNewBook = false
+      this.showForm = true
     },
     // 入力フォームクローズ
     closeForm () {
-      this.$refs.form.reset()
       this.showForm = false
     },
-    // 書籍削除ダイアログ表示
-    openDeleteDialog (item) {
-      this.dialogDelete = true
-      this.selectedBookId = item.bookId
-      this.selectedTitle = item.title
-    },
-    // 書籍削除決定
-    async deleteItemConfirm () {
-      this.closeDelete()
-      this.overlay = true
-      try {
-        await this.deleteItem() // 書籍削除
-        this.items = await this.readBookTableAll() // 書籍一覧更新
-      } catch (e) {
-        this.notifyFailure()
-      }
-      this.overlay = false
-    },
-    // 削除ダイアログクローズ
-    closeDelete () {
-      this.dialogDelete = false
-    },
-    // 書籍編集フォーム表示
-    openEditForm (item) {
-      this.selectedBookId = item.bookId
-      this.formTitle = item.title
-      this.selectedFormGenreList = item.genreId
-      this.pickerDate = item.purchaseDate
-      this.formPurchaseName = item.purchaseName
-      this.formReview = item.review
-      this.showForm = true
-    },
-    // 書籍情報更新サブミット
-    async submitUpdateItem () {
-      this.overlay = true
-      try {
-        await this.updateItem() // 書籍情報更新
-        this.items = await this.readBookTableAll() // 書籍一覧更新
-      } catch (e) {
-        this.notifyFailure()
-      }
-      this.overlay = false
-      this.closeForm()
-    },
-    // 書籍新規登録と既存書籍更新の処理を振り分けるハンドラー
-    handlerInsertOrUpdate () {
-      if (this.selectedBookId === 0) {
-        this.submitInsertNewItem() // 新規登録
-      } else {
-        this.submitUpdateItem() // 既存更新
-      }
-    },
-    // 書籍レビューダイアログオープン
-    showReview (item) {
-      this.formTitle = item.title
-      this.formReview = item.review
-      this.dialogReview = true
-    },
-    // 書籍レビューダイアログクローズ
-    closeReview () {
-      this.dialogReview = false
-      this.formTitle = ''
-      this.formReview = ''
-    },
     // 書籍を検索
-    async searchBook () {
+    async searchBook (
+      searchTitle,
+      searchGenre) {
+      this.searchTitle = searchTitle
+      this.searchGenre = searchGenre
       if (!this.searchTitle && !this.searchGenre) {
         alert('検索条件が指定されていません')
       } else {
@@ -471,8 +177,6 @@ export default {
     },
     // 検索条件クリア
     async searchClear () {
-      this.searchTitle = ''
-      this.searchGenre = ''
       // 検索結果を表示中の場合は全件表示に戻す
       if (this.isFilteredView) {
         try {
@@ -493,30 +197,3 @@ export default {
   }
 }
 </script>
-
-<style>
-.v-data-table th {
-    background: #90CAF9;
-}
-#overlay{
-  /* 要素を重ねた時の順番 */
-  z-index:1;
-  /* 画面全体を覆う設定 */
-  position:fixed;
-  top:0;
-  left:0;
-  width:100%;
-  height:100%;
-  background-color:rgba(0,0,0,0.5);
-  /* 画面の中央に要素を表示させる設定 */
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-#formWindow{
-  z-index:2;
-  width:30%;
-  padding: 1em;
-  background:#90CAF9;
-}
-</style>
